pyproject.toml
[project]
name = "kachi"
version = "0.1.0"
description = "Make a protected link unsafe"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0,<7.0"]
license = "MIT"
authors = [{ name = "Manabu Niseki", email = "manabu.niseki@gmail.com" }]

[build-system]
requires = ["hatchling>=1.26", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.hooks.version]
path = "src/kachi/_version.py"
template = '''
version = "{version}"
'''

[dependency-groups]
dev = [
  "mypy>=1.19.1",
  "prek>=0.3.2",
  "pytest-pretty>=1.3.0",
  "pytest-randomly>=4.0.1",
  "pytest>=9.0.2",
  "ruff>=0.15.0",
  "types-pyyaml>=6.0.12.20250915",
  "uv-sort>=0.7.0",
]

[tool.ruff.lint]
select = [
  "ARG",  # flake8-unused-arguments
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "FA",   # flake8-future-annotations
  "I",    # isort
  "N",    # pep8-naming
  "PERF", # Perflint
  "PT",   # flake8-pytest-style
  "RET",  # flake8-return
  "RUF",  # Ruff-specific rules
  "SIM",  # flake8-simplify
  "T20",  # flake8-print
  "UP",   # pyupgrade
  "W",    # pycodestyle warnings
]
ignore = ["E501"]

[tool.mypy]
ignore_missing_imports = true
